pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.2.2",
    "daphne>=4.1.2",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "paho-mqtt>=2.1.0",
    "psycopg2-binary>=2.9.10",
]
